requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
Plaid-Python==8.6.0
SQLAlchemy==1.4.41
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
Flask-Login==0.6.2
Flask-Cors==3.0.10
gunicorn==20.1.0
werkzeug==2.2.2
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
blinker==1.5
